Remove endpoint defaults from JS layer
In the case that the hosted version of Bugsnag is used, the defaults are
inferred from the native layer. In the case that an on-premise version
is used, these values are overridden. This change avoids the case where
the endpoint(s) are overridden in the native layer and reset to the
defaults by the JS layer.