fix: Define sessions as starting when JS layer is initialized
Disables the native logic for automatically reporting the first session
with a call from startWithOptions. This averts issues where the native
release stage does not match the eventual actual release stage set on
the JS side. This change assumes that React Native components are the
primary way to interact with the application, and when this is not the
case, a different session tracking setup should be used instead of the
automatic session detection.

Mentioned in bugsnag-cocoa#295