fix: add proguard consumer rules to retain JavascriptException class
Any exception of the type com.facebook.react.common.JavascriptException is ignored by the notifier.
This is because the error is captured in the Android + JS layers, and the JS error contains more
relevant information, so the Android exception is discarded to avoid duplicates. If ProGuard is
enabled, the class name will be obfuscated and therefore the ignore class will not work. This adds a
keep rule for this class, to prevent duplicate errors.